+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Fixed
 
 * Regression in [HexaPDF::Layout::ListBox] that leads to missing markers
+* [HexaPDF::Content::CanvasComposer#draw_box] to handle truncated boxes
 
 
 ## 0.44.0 - 2024-06-05